WIP: Avoid redundant motor configuration
Refactor the motor configuration structure to avoid redundancy:

Instead of having duplicate parameter defines in all `user*_j1/5.h`
files, add a separate header `motors.h` in which motor parameters are
stored using const instances of a struct `MOTORCONFIG_MotorData_t`.

In the "user" files, the motor to be used is defined by a
`USER_MotorData` define which is set to the corresponding constant.
This is then used to fill the old parameter defines with the
corresponding values (to stay compatible with all the MotorWare code
that expects those defines).